fix(core): widen cpu_cycles from u32 to u64 to prevent overflow
Change cpu_cycles type from u32 to u64 throughout the measurement stack
to prevent overflow in benchmarks exceeding 1 second at 4GHz.

Changes:
- Sample struct: cpu_cycles field widened to u64
- Timer::stop(): returns u64 cycles instead of u32
- Bencher: accumulate_sample() and internal accumulators use u64
- IPC messages: Sample serialization updated for 32-byte layout
- Statistics: CyclesStatistics uses u64 for min/max cycles
- Reports: BenchmarkMetrics updated to match

The previous u32 representation (4 billion cycles) was sufficient for
sub-second measurements but caused truncation in longer benchmarks.
Full 64-bit representation ensures accurate cycle counts for benchmarks
of any duration.

+++ fluxbench-cli/src/executor/execution.rs
@@ -81,7 +81,7 @@ pub struct BenchExecutionResult {
     pub status: BenchmarkStatus,
     pub samples: Vec<f64>,
     /// CPU cycles per sample (parallel with samples)
-    pub cpu_cycles: Vec<u32>,
+    pub cpu_cycles: Vec<u64>,
     pub alloc_bytes: u64,
     pub alloc_count: u64,
     pub duration_ns: u64,
@@ -155,7 +155,7 @@ impl Executor {
                     .collect();
 
                 // Extract CPU cycles (parallel array with samples)
-                let cpu_cycles: Vec<u32> =
+                let cpu_cycles: Vec<u64> =
                     bench_result.samples.iter().map(|s| s.cpu_cycles).collect();
 
                 // Sum allocations
@@ -326,7 +326,7 @@ impl IsolatedExecutor {
             .collect();
 
         // Extract CPU cycles
-        let cpu_cycles: Vec<u32> = ipc_result.samples.iter().map(|s| s.cpu_cycles).collect();
+        let cpu_cycles: Vec<u64> = ipc_result.samples.iter().map(|s| s.cpu_cycles).collect();
 
         // Sum allocations
         let alloc_bytes: u64 = ipc_result.samples.iter().map(|s| s.alloc_bytes).sum();
